fl/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.7.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py38haa95532_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2022.07.19=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.6.15=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
colorama=0.4.5=py38haa95532_0
contourpy=1.0.5=pypi_0
cpuonly=2.0=0
cryptography=37.0.1=py38h21b164f_0
cycler=0.11.0=pypi_0
d2l=0.17.5=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py38h2bbff1b_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py38hc60d5dd_0
idna=2.10=pypi_0
importlib-metadata=4.11.3=py38haa95532_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py38haa95532_0
ipython=8.4.0=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=py38haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
jsonschema=4.4.0=py38haa95532_0
jupyter=1.0.0=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter_client=7.2.2=py38haa95532_0
jupyter_core=4.10.0=py38haa95532_0
jupyterlab-widgets=3.0.3=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libdeflate=1.8=h2bbff1b_5
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h8a3f274_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py38haa95532_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=5.1.0=py38h2bbff1b_2
nb_conda=2.2.1=py38_1
nb_conda_kernels=2.3.1=py38haa95532_0
nbclient=0.5.13=py38haa95532_0
nbconvert=6.4.4=py38haa95532_0
nbformat=5.3.0=py38haa95532_0
nest-asyncio=1.5.5=py38haa95532_0
netaddr=0.8.0=pypi_0
notebook=6.4.12=py38haa95532_0
numpy=1.21.5=pypi_0
oauthlib=3.2.0=pyhd3eb1b0_1
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.2.4=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hdc2b20a_1
pip=22.1.2=py38haa95532_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py38hd77b12b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.9.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py38haa95532_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38haa95532_0
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
pytorch=1.12.1=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pywin32=302=py38h2bbff1b_2
pywinpty=2.0.2=py38h5da7b33_0
pyzmq=23.2.0=py38hd77b12b_0
qtconsole=5.3.2=pypi_0
qtpy=2.2.1=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.8.0=py38haa95532_0
tensorboard-data-server=0.6.0=py38haa95532_0
tensorboard-plugin-wit=1.8.1=py38haa95532_0
terminado=0.13.1=py38haa95532_0
testpath=0.6.0=py38haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pypi_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.12.1=py38_cpu
torchvision=0.13.1=py38_cpu
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py38haa95532_0
typing_extensions=4.3.0=py38haa95532_0
urllib3=1.26.11=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
xz=5.2.5=h8cc25b3_1
yarl=1.8.1=py38h2bbff1b_0
zeromq=4.3.4=hd77b12b_0
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
